refactored the testcase code
