+++ mockingServer/data/vcsa_output_comparison.yaml
