+++ mockingServer/modules/RequestHandler.py
@@ -5,7 +5,7 @@ import json
 
 class RequestHandler(MethodView):
     def __init__(self):
-        self.session_id = {'value': '779b8ede-8332-11eb-9581-3c58c27e75a6'}
+        self.session_id = {'value': '779b8ede-1337-11eb-9581-3c58c27e75a6'}
 
     def check_session_id(self):
         if request.headers.environ['HTTP_VMWARE_API_SESSION_ID'] != self.session_id['value']:
